manually close leftover file descriptors in child
When we fork and exec the child process, if the parent process has opened any
file descriptors (other than stdin, stdout, stderr) and hasn't flagged them as
FD_CLOEXEC, they won't get closed. Consequently, if the parent is also doing
something complicated with file descriptors, forking and execing the child can
cause hangs on shutdown. This solution appears to work on Linux. It may work on
OS X as well, but it probably won't work on Windows, unfortunately.